Created tests for animals
