+++ users/routes.spec.js
@@ -1,3 +1,6 @@
+const request = require("supertest");
+const server = require("../server.js");
+
 describe("user routes", () => {
   it("should respond with 401 created & json on /api/users/register", async () => {
     const res = await request(server).post("/api/users/register", {
@@ -31,21 +34,6 @@ describe("user routes", () => {
     const res = await request(server).post("/api/users/login", {
       username: "badLogin"
     });
-    expect(res.type).toBe("application/json");
-    expect(res.status).toBe(400);
-  });
-});
-
-describe("animal routes", () => {
-  it("should return json and 200 OK", async () => {
-    const headers = { headers: { Authorization: "string" } };
-    const res = await request(server).get("/api/animals", headers);
-    expect(res.type).toBe("application/json");
-    expect(res.status).toBe(200);
-  });
-
-  it("should return 400 without a header", async () => {
-    const res = await request(server).get("/api/animals");
     expect(res.type).toBe("application/json");
     expect(res.status).toBe(400);
   });
